finished reservation system Admin page has been made Overview page has been made

+++ src/main/java/ServletDeleteReservations.java
@@ -13,26 +13,36 @@ public class ServletDeleteReservations extends HttpServlet{
         String email = req.getParameter("email");
         String reservationId = req.getParameter("reservationId");
         String workspaceId = req.getParameter("workspaceId");
+        String hostId = req.getParameter("employeeId");
+        String inviteeAccepted = req.getParameter("inviteeaccepted");
 
-        ResultSet resultSetId = DatabaseManager.getResultsFromQuery("select firstname, lastname from employeeTable where emailAddress='"+email+"'");
+        ResultSet resultSetId = DatabaseManager.getResultsFromQuery("select firstname, lastname, employeeID from employeeTable where emailAddress='"+email+"'");
 
-        String resultNameString = "";
-        String resultSurnameString = "";
+        String firstName = "";
+        String lastName = "";
+        String id = "";
 
         try {
             if(resultSetId.next()) {
-                resultNameString = resultSetId.getString(1);
-                resultSurnameString = resultSetId.getString(2);
-                System.out.println(resultNameString+" "+resultSurnameString);
+                firstName = resultSetId.getString(1);
+                lastName = resultSetId.getString(2);
+                req.setAttribute("Id", resultSetId.getString(3));
+                id = resultSetId.getString(3);
+                System.out.println(firstName+" "+lastName);
             }
         } catch (SQLException throwables) {
             System.out.println("employee name is missing");
             throwables.printStackTrace();
         }
 
-        DatabaseManager.executeSQLstatement("delete from invitationtable where reservationid='" + reservationId + "'");
-        DatabaseManager.executeSQLstatement("delete from reservationtable where reservationid='" + reservationId + "'");
-        DatabaseManager.executeSQLstatement("delete from workspacetable where workspaceid='" + workspaceId + "'");
+        if (id.equals(hostId)) {
+            DatabaseManager.executeSQLstatement("delete from invitationtable where reservationid='" + reservationId + "'");
+            DatabaseManager.executeSQLstatement("delete from reservationtable where reservationid='" + reservationId + "'");
+            DatabaseManager.executeSQLstatement("delete from workspacetable where workspaceid='" + workspaceId + "'");
+        }
+        else {
+            DatabaseManager.executeSQLstatement("update invitationtable set inviteeaccepted = array_remove(inviteeaccepted, '" + firstName+" "+lastName + "')" + "where '" + firstName+" "+lastName + "' = any(inviteeaccepted) and reservationid='"+reservationId+"'");
+        }
 
         System.out.println(invitedby + " invitedby");
         System.out.println(email + " email");
